fix(week2-review): address all review findings
[P1] u128 balance consistency:
- Account.balance: u64 -> u128
- StateDB.balance(): u64 -> u128
- StateDB.decrease_balance/increase_balance: u64 -> u128
- Error::InsufficientBalance: u64 -> u128
- Removed all value cast from u128 to u64 in executor

[P1] Atomic block execution:
- execute_block now clones working state, executes all txs,
  only commits on success
- On any tx failure, original state remains unchanged
- Added test_block_atomic_on_failure

[P1] Receipt root from receipt content:
- Added Receipt::receipt_hash() that hashes actual receipt fields
- receipt_root now uses receipt_hash() not tx hash
- Receipt hash includes: tx_hash || success || gas_used || logs_hash

[P2] clippy --all-targets fixes:
- Fixed needless borrows in types.rs receipt_hash()
- Fixed &&addr/&&[0u8; 20] in state.rs tests

[P1] ExecutionOutcome struct:
- execute_block returns ExecutionOutcome { receipts, tx_root, receipt_root, state_root, gas_used }
- Added execute_tx internal helper for atomic execution

Tests: 24 passed
Verification:
- cargo fmt --check: PASS
- cargo clippy --all-targets -D warnings: PASS
- cargo test: 24 passed

+++ src/error.rs
@@ -8,7 +8,7 @@ pub enum Error {
     InvalidTransaction(String),
 
     #[error("insufficient balance: have {have}, need {need}")]
-    InsufficientBalance { have: u64, need: u64 },
+    InsufficientBalance { have: u128, need: u128 },
 
     #[error("invalid nonce: expected {expected}, got {got}")]
     InvalidNonce { expected: u64, got: u64 },
